fix: move modules and lint files
