+++ src/execute_system.c
@@ -6,57 +6,11 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/08 02:27:44 by vwildner          #+#    #+#             */
-/*   Updated: 2022/05/12 01:30:17 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/05/12 02:17:21 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-int	ft_listlen(t_list **list)
-{
-	t_list	*tmp;
-	int		len;
-
-	len = 0;
-	tmp = *list;
-	while (tmp)
-	{
-		tmp = tmp->next;
-		len++;
-	}
-	return (len);
-}
-
-char	**to_array(t_list **list)
-{
-	char	**array;
-	t_list	*tmp;
-	int		i;
-	int		size;
-
-	i = 0;
-	size = ft_listlen(list);
-	tmp = *list;
-	array = (char **)malloc(sizeof(char *) * (size + 1));
-	while (tmp)
-	{
-		array[i++] = tmp->content;
-		tmp = tmp->next;
-	}
-	return (array);
-}
-
-char	*solve_absolute_path(t_command *cmd)
-{
-	char	*first_arg;
-	char	*all_paths;
-
-	first_arg = cmd->argv[0];
-	if (*first_arg == '/' || *first_arg == '.')
-		return (first_arg);
-	all_paths = ms_getenv(cmd->envp, "PATH");
-	return (get_abspath(cmd, first_arg, all_paths));
-}
 
 void	read_file(char *filename)
 {
@@ -74,8 +28,9 @@ void	read_file(char *filename)
 
 void	clear_first_arg(t_command *cmd, int first_arg_pos)
 {
-	int	i;
+	int		i;
 	char	**tmp;
+
 	i = 0;
 	while (cmd->argv[i])
 		i++;
@@ -93,7 +48,6 @@ void	clear_first_arg(t_command *cmd, int first_arg_pos)
 
 int	handle_first_arg(t_command *cmd)
 {
-
 	int		first_arg_pos;
 	char	*filename;
 
@@ -113,7 +67,7 @@ int	handle_first_arg(t_command *cmd)
 	return (0);
 }
 
-static void	execute_child_command(t_command *cmd)
+void	execute_child_command(t_command *cmd)
 {
 	char	*abspath;
 	char	**compat_envp;
