+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/16 18:54:55 by itaureli          #+#    #+#             */
-/*   Updated: 2022/05/11 18:26:46 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/05/12 02:13:44 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,19 +52,6 @@ char	*help_slash_merge(char const *origin, char const *other)
 	return (merged);
 }
 
-void	print_err_msg(char *command, char *msg)
-{
-	char	*err_msg;
-
-	err_msg = ft_strdup("bash: ");
-	err_msg = ft_strjoin(err_msg, command);
-	err_msg = ft_strjoin(err_msg, ": ");
-	err_msg = ft_strjoin(err_msg, msg);
-	err_msg = ft_strjoin(err_msg, "\n");
-	write(STDERR_FILENO, err_msg, ft_strlen(err_msg));
-	free(err_msg);
-}
-
 char	*get_abspath(t_command *cmd, char *command, const char *path)
 {
 	char	*file;
@@ -92,4 +79,16 @@ char	*get_abspath(t_command *cmd, char *command, const char *path)
 	cmd->status = 127;
 	exit(EXIT_FAILURE);
 	return (NULL);
+}
+
+char	*solve_absolute_path(t_command *cmd)
+{
+	char	*first_arg;
+	char	*all_paths;
+
+	first_arg = cmd->argv[0];
+	if (*first_arg == '/' || *first_arg == '.')
+		return (first_arg);
+	all_paths = ms_getenv(cmd->envp, "PATH");
+	return (get_abspath(cmd, first_arg, all_paths));
 }
