+++ src/utils2.c
@@ -1,0 +1,60 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils2.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/08 02:27:44 by vwildner          #+#    #+#             */
+/*   Updated: 2022/05/12 02:13:06 by vwildner         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+void	print_err_msg(char *command, char *msg)
+{
+	char	*err_msg;
+
+	err_msg = ft_strdup("bash: ");
+	err_msg = ft_strjoin(err_msg, command);
+	err_msg = ft_strjoin(err_msg, ": ");
+	err_msg = ft_strjoin(err_msg, msg);
+	err_msg = ft_strjoin(err_msg, "\n");
+	write(STDERR_FILENO, err_msg, ft_strlen(err_msg));
+	free(err_msg);
+}
+
+int	ft_listlen(t_list **list)
+{
+	t_list	*tmp;
+	int		len;
+
+	len = 0;
+	tmp = *list;
+	while (tmp)
+	{
+		tmp = tmp->next;
+		len++;
+	}
+	return (len);
+}
+
+char	**to_array(t_list **list)
+{
+	char	**array;
+	t_list	*tmp;
+	int		i;
+	int		size;
+
+	i = 0;
+	size = ft_listlen(list);
+	tmp = *list;
+	array = (char **)malloc(sizeof(char *) * (size + 1));
+	while (tmp)
+	{
+		array[i++] = tmp->content;
+		tmp = tmp->next;
+	}
+	return (array);
+}
